# Show a JSON report file content when RSpec fails during a dry run