--- a/lib/knapsack_pro/test_case_detectors/rspec_test_example_detector.rb
+++ b/lib/knapsack_pro/test_case_detectors/rspec_test_example_detector.rb
@@ -48,7 +48,7 @@ module KnapsackPro
 
           KnapsackPro.logger.error('-'*10 + ' END of actionable error message ' + '-'*50)
 
-          raise 'There was a problem while generating test examples for the slow test files. Please read actionable error message above.'
+          raise 'There was a problem while generating test examples for the slow test files. Please read the actionable error message above.'
         end
       end