--- a/spec/knapsack_pro/test_case_detectors/rspec_test_example_detector_spec.rb
+++ b/spec/knapsack_pro/test_case_detectors/rspec_test_example_detector_spec.rb
@@ -76,7 +76,7 @@ describe KnapsackPro::TestCaseDetectors::RSpecTestExampleDetector do
           end
 
           it do
-            expect { subject }.to raise_error(RuntimeError, 'There was a problem while generating test examples for the slow test files. Please read actionable error message above.')
+            expect { subject }.to raise_error(RuntimeError, 'There was a problem while generating test examples for the slow test files. Please read the actionable error message above.')
           end
         end
       end